- Fix Profile loading - Fix mouse zoom and info - Fix Preferences button - Fix low precision output

+++ plot_script/measurement_data_structure.py
@@ -929,7 +929,7 @@ class MeasurementData(object):
 
   def export(self, file_name, print_info=True, seperator=' ',
              xfrom=None, xto=None, only_fitted_columns=False,
-             format_string="%g"):
+             format_string="%.15g"):
     '''
       Write data in text file.
       
@@ -1065,7 +1065,7 @@ class MeasurementData(object):
     srt_run2=numpy.lexsort(keys=(ds1, ds2))
     return srt_run2
 
-  def string_from_data_matrix(self, seperator, data, split_indices, format_string="%g"):
+  def string_from_data_matrix(self, seperator, data, split_indices, format_string="%.15g"):
     '''
       Create a string that can be written to a file from a given data matrix.
       The function may look quite strange because of a lot of optimization,
@@ -2583,7 +2583,8 @@ class PhysicalProperty(numpy.ndarray):
 
   unit_save=True #: if true changes units after arithmetic operation and checks if correct
 
-  def __new__(cls, dimension_in, unit_in, input_data=[], input_error=None, unit_save=True):
+  def __new__(cls, dimension_in, unit_in, input_data=[], input_error=None, unit_save=True,
+              dtype=numpy.float32):
     '''
       Class constructor when explcidly called.
       
@@ -2591,14 +2592,14 @@ class PhysicalProperty(numpy.ndarray):
       :param unit_in: String with unit for that instance
     '''
     #obj=numpy.ndarray.__new__(cls, len(input_data), dtype=numpy.float32)
-    obj=numpy.asarray(input_data, dtype=numpy.float32).view(cls).copy()
+    obj=numpy.asarray(input_data, dtype=dtype).view(cls).copy()
     obj.unit=PhysicalUnit(unit_in)
     obj.dimension=dimension_in
     obj.unit_save=True
     if input_error is not None:
       if len(input_error)!=len(input_data):
         raise ValueError, 'shape mismatch: error and data have different lengths'
-      obj.error=numpy.array(input_error, dtype=numpy.float32)
+      obj.error=numpy.array(input_error, dtype=dtype)
     obj.__setslice__(0, len(input_data), input_data)
     return obj
